requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
django-cors-headers==3.13.0
django-countries==7.3.2
django-createsuperuserwithpassword==2.0.0
django-phonenumber-field==6.1.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
Faker==13.13.0
gunicorn==20.1.0
Pillow==9.1.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
whitenoise==6.2.0
